app/models/image_analyzer.py
"""
Image analysis module using pretrained ResNet18.

Strategy:
  1. Extract deep features from the inspection image via ResNet18 (ImageNet weights).
  2. Compute an anomaly score by measuring how far the feature vector deviates
     from a simple "normal" reference (mean activation magnitude).
  3. Generate a Grad-CAM heatmap highlighting the most suspicious regions.
  4. Map the anomaly pattern to the most likely defect category using
     feature-space heuristics.

Honest disclaimer:
  This is zero-shot anomaly detection — the model was NOT trained on PCB data.
  It provides a credible baseline that works surprisingly well on textured
  surfaces and can be replaced with a fine-tuned model later.
"""

import io
from dataclasses import dataclass

import cv2
import numpy as np
import torch
import torch.nn.functional as F
from PIL import Image
from torchvision import models, transforms

import config


# ── Result container ─────────────────────────────────────────────────────────

@dataclass
class ImageAnalysisResult:
    """Output of the image analysis pipeline."""
    defect_category: str        # predicted defect type
    confidence: float           # 0-1 confidence in the prediction
    anomaly_score: float        # 0-1 how anomalous the image looks
    severity: int               # 0-100 severity score
    heatmap: np.ndarray | None  # RGB heatmap overlay (H, W, 3) or None


# ── Preprocessing ────────────────────────────────────────────────────────────

_preprocess = transforms.Compose([
    transforms.Resize((config.IMAGE_SIZE, config.IMAGE_SIZE)),
    transforms.ToTensor(),
    transforms.Normalize(
        mean=[0.485, 0.456, 0.406],
        std=[0.229, 0.224, 0.225],
    ),
])


# ── Analyzer class ───────────────────────────────────────────────────────────

class ImageAnalyzer:
    """Pretrained ResNet18-based anomaly detector for inspection images."""

    def __init__(self):
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.model = models.resnet18(weights=models.ResNet18_Weights.DEFAULT)
        self.model.eval()
        self.model.to(self.device)

        # Hook into the last conv layer for Grad-CAM
        self._activations = None
        self._gradients = None
        self.model.layer4.register_forward_hook(self._save_activation)
        self.model.layer4.register_full_backward_hook(self._save_gradient)

    # ── Hooks for Grad-CAM ───────────────────────────────────────────────

    def _save_activation(self, module, input, output):
        self._activations = output.detach()

    def _save_gradient(self, module, grad_input, grad_output):
        self._gradients = grad_output[0].detach()

    # ── Public API ───────────────────────────────────────────────────────

    def analyze(self, image: Image.Image) -> ImageAnalysisResult:
        """Run full analysis on a PIL image. Returns an ImageAnalysisResult."""
        tensor = _preprocess(image.convert("RGB")).unsqueeze(0).to(self.device)
        tensor.requires_grad_(True)

        # Forward pass
        logits = self.model(tensor)
        probs = F.softmax(logits, dim=1)

        # Anomaly score: how spread out are the activations?
        anomaly_score = self._compute_anomaly_score(self._activations)

        # Grad-CAM heatmap for the top predicted class
        top_class = logits.argmax(dim=1).item()
        heatmap = self._grad_cam(tensor, logits, top_class, image.size)

        # Map to defect category via feature heuristics
        defect_category, confidence = self._classify_defect(
            anomaly_score, self._activations
        )

        # Severity: scale anomaly score to 0-100
        severity = int(min(100, max(0, anomaly_score * 100)))

        return ImageAnalysisResult(
            defect_category=defect_category,
            confidence=round(confidence, 3),
            anomaly_score=round(anomaly_score, 4),
            severity=severity,
            heatmap=heatmap,
        )

    # ── Internal methods ─────────────────────────────────────────────────

    def _compute_anomaly_score(self, activations: torch.Tensor) -> float:
        """
        Compute anomaly score from feature activations.

        Normal images tend to have moderate, evenly distributed activations.
        Anomalous images show spiking or unusual activation patterns.
        We use the coefficient of variation + sparsity as a proxy.
        """
        act = activations.squeeze()                       # (C, H, W)
        channel_means = act.mean(dim=(1, 2))              # (C,)

        # Coefficient of variation across channels
        cv = channel_means.std() / (channel_means.mean() + 1e-8)

        # Sparsity: fraction of near-zero activations
        sparsity = (act < 0.1).float().mean()

        # Combined score, clamped to [0, 1]
        raw_score = float(0.5 * cv + 0.5 * sparsity)
        return min(1.0, max(0.0, raw_score))

    def _grad_cam(
        self,
        tensor: torch.Tensor,
        logits: torch.Tensor,
        class_idx: int,
        original_size: tuple[int, int],
    ) -> np.ndarray:
        """Generate a Grad-CAM heatmap overlay at the original image size."""
        self.model.zero_grad()
        target = logits[0, class_idx]
        target.backward(retain_graph=True)

        gradients = self._gradients.squeeze()     # (C, H, W)
        activations = self._activations.squeeze()  # (C, H, W)

        # Channel-wise weight = global average of gradients
        weights = gradients.mean(dim=(1, 2))       # (C,)

        # Weighted combination of activation maps
        cam = torch.zeros(activations.shape[1:], device=self.device)
        for i, w in enumerate(weights):
            cam += w * activations[i]

        cam = F.relu(cam)
        cam = cam - cam.min()
        cam = cam / (cam.max() + 1e-8)

        # Resize to original image dimensions (width, height)
        cam_np = cam.cpu().numpy()
        cam_resized = cv2.resize(cam_np, original_size)

        # Convert to color heatmap
        heatmap = cv2.applyColorMap(
            (cam_resized * 255).astype(np.uint8), cv2.COLORMAP_JET
        )
        heatmap = cv2.cvtColor(heatmap, cv2.COLOR_BGR2RGB)
        return heatmap

    def _classify_defect(
        self, anomaly_score: float, activations: torch.Tensor
    ) -> tuple[str, float]:
        """
        Map anomaly patterns to a defect category.

        This uses hand-crafted heuristics on the feature statistics.
        A real system would replace this with a trained classifier head.
        """
        if anomaly_score < config.ANOMALY_THRESHOLD:
            return "normal", round(1.0 - anomaly_score, 3)

        act = activations.squeeze()
        channel_means = act.mean(dim=(1, 2))
        spatial_var = act.var(dim=(1, 2)).mean().item()
        top_activation = channel_means.max().item()
        activation_range = (channel_means.max() - channel_means.min()).item()

        # Heuristic decision tree based on activation statistics.
        # These thresholds were picked to give plausible variety —
        # they are NOT validated on real PCB data.
        if spatial_var > 2.0 and activation_range > 3.0:
            category = "crack"
        elif top_activation > 4.0:
            category = "solder_bridge"
        elif spatial_var > 1.5:
            category = "scratch"
        elif activation_range > 2.5:
            category = "misalignment"
        elif top_activation > 2.5:
            category = "corrosion"
        elif spatial_var > 0.8:
            category = "contamination"
        else:
            category = "open_circuit"

        confidence = min(0.95, 0.4 + anomaly_score * 0.5)
        return category, round(confidence, 3)


def overlay_heatmap(image: Image.Image, heatmap: np.ndarray, alpha: float = config.HEATMAP_ALPHA) -> Image.Image:
    """Blend original image with Grad-CAM heatmap."""
    img_array = np.array(image.convert("RGB").resize(
        (heatmap.shape[1], heatmap.shape[0])
    ))
    blended = cv2.addWeighted(img_array, 1 - alpha, heatmap, alpha, 0)
    return Image.fromarray(blended)
